feat: add pdf signature
Add the ability to add a signature to a PDF document.
This includes functionality to pick a PDF, display it page by page,
and position a signature image using drag and drop.
The user can toggle the signature visibility.
Also, a start to save as PDF is added.
Remove unused import and comment.